requirements.txt
discord
PyNaCl
python-dotenv
python-ffmpeg
youtube_dl
youtube-search-python
pytube